FINAL_Back_Template_Karate

+++ Karate_FINAL/view_students.php
@@ -1,10 +1,10 @@
-<?php 
-
-$page_title = 'Students';
-include ('manage.html');
+<?php
+require_once ('./config_inc.php'); 
+include_once(HEADER);
+require_once ('./upload_image.php');
 echo '<h1>Students</h1>';
 
-require ('connector.php');
+require_once (CONNECTOR);
 
 // Number of records to show per page:
 $display = 10;
@@ -14,7 +14,7 @@ if (isset($_GET['p']) && is_numeric($_GET['p'])) { // Already been determined.
 	$pages = $_GET['p'];
 } else { // Need to determine.
  	// Count the number of records:
-	$q = "SELECT COUNT(ID) FROM to2446992_kv2";
+	$q = "SELECT COUNT(ID) FROM to2446992_Karate_Student_App";
 	$r = @mysqli_query ($database, $q);
 	$row = @mysqli_fetch_array ($r, MYSQLI_NUM);
 	$records = $row[0];
@@ -39,6 +39,9 @@ $sort = (isset($_GET['sort'])) ? $_GET['sort'] : 'rd';
 
 // Determine the sorting order:
 switch ($sort) {
+	case 'ID':
+		$order_by = 'ID ASC';
+		break;
 	case 'ln':
 		$order_by = 'LastName ASC';
 		break;
@@ -48,6 +51,9 @@ switch ($sort) {
 	case 'rd':
 		$order_by = 'RegDate ASC';
 		break;
+	case 'Photo':
+		$order_by = 'Photo ASC';
+		break;
 	default:
 		$order_by = 'RegDate ASC';
 		$sort = 'rd';
@@ -55,28 +61,36 @@ switch ($sort) {
 }
 	
 // Define the query:
-$q = "SELECT LastName, FirstName, DATE_FORMAT,Photo(RegDate, '%M %d, %Y') AS dr,  FROM to2446992_kv2 ORDER BY $order_by LIMIT $start, $display";		
-$r = @mysqli_query ($database, $q); // Run the query.
+$q = "SELECT LastName, FirstName, DATE_FORMAT(RegDate, '%M %d, %Y')as dr, ID, Photo FROM to2446992_Karate_Student_App";
 
-// Table header:
-echo '<table align="center" cellspacing="0" cellpadding="5" width="75%">
-<tr>
-	<td align="left"><b>Edit</b></td>
-	<td align="left"><b>Delete</b></td>
+$r = @mysqli_query($database, $q); // Run the query.
 
-	<td align="left"><b><a href="view_students.php?sort=ln">Last Name</a></b></td>
-	<td align="left"><b><a href="view_students.php?sort=fn">First Name</a></b></td>
-	<td align="left"><b><a href="view_students.php?sort=rd">Date Registered</a></b></td>
-</tr>
-';
+    	// Table header:
+			echo '<table align="center" cellspacing="0" cellpadding="5" width="75%">
+			<tr>
+				<td align="left"><b>Edit</b></td>
+				<td align="left"><b>Delete</b></td>
+				<td align="left"><b><a href="view_students.php?sort=ID">ID</a></b></td>
+				<td align="left"><b><a href="view_students.php?sort=Photo">Photo</a></b></td>
+				<td align="left"><b><a href="view_students.php?sort=ln">Last Name</a></b></td>
+				<td align="left"><b><a href="view_students.php?sort=fn">First Name</a></b></td>
+				<td align="left"><b><a href="view_students.php?sort=rd">Date Registered</a></b></td>
+			</tr>
+			';
 
-// Fetch and print all the records....
-$bg = '#eeeeee'; 
-while ($row = mysqli_fetch_array($r, MYSQLI_ASSOC)) {
-	$bg = ($bg=='#eeeeee' ? '#ffffff' : '#eeeeee');
+			// Fetch and print all the records....
+			$bg = '#000000'; 
+if($r === FALSE) {
+    die(mysqli_error($database)); // TODO: better error handling
+}
+while ($row = mysqli_fetch_array($r, MYSQLI_ASSOC))
+
+ {
+	$bg = ($bg=='#000000' ? '#ffffff' : '#000000');
 		echo '<tr bgcolor="' . $bg . '">
 		<td align="left"><a href="edit_user.php?id=' . $row['ID'] . '">Edit</a></td>
 		<td align="left"><a href="delete_user.php?id=' . $row['ID'] . '">Delete</a></td>
+		<td align="left">' . $row['ID'] . '</td>
 		<td align="left">' . $row['LastName'] . '</td>
 		<td align="left">' . $row['Photo'] . '</td>
 		<td align="left">' . $row['FirstName'] . '</td>
@@ -117,6 +131,5 @@ if ($pages > 1) {
 	echo '</p>'; // Close the paragraph.
 	
 } // End of links section.
-	
-include ('footer.html');
+
 ?>
